增加了大量和native相关的代码

+++ src/page.mp.ts
@@ -238,15 +238,18 @@ const oakBehavior = Behavior<
         },
 
         save(key: string, item: any) {
-            this.features.localStorage.save(key, item);
+            return this.features.localStorage.save(key, item);
         },
 
         load(key: string) {
             return this.features.localStorage.load(key);
         },
 
-        clear() {
-            this.features.localStorage.clear();
+        clear(key?: string) {
+            if (key) {
+                return this.features.localStorage.remove(key);
+            }
+            return this.features.localStorage.clear();
         },
 
         setNotification(data: NotificationProps) {
